Avances modo de juego draft

+++ app/src/main/java/com/nbastatsquiz/beans/DraftPick.java
@@ -4,11 +4,11 @@ import java.io.Serializable;
 
 public class DraftPick implements Serializable {
 
-    Double PERSON_ID, ROUND_NUMBER, ROUND_PICK, OVERALL_PICK, TEAM_ID;
+    int PERSON_ID, ROUND_NUMBER, ROUND_PICK, OVERALL_PICK, TEAM_ID;
     String PLAYER_NAME, SEASON, DRAFT_TYPE, TEAM_CITY, TEAM_NAME, TEAM_ABBREVIATION, ORGANIZATION, ORGANIZATION_TYPE;
 
 
-    public DraftPick(Double PERSON_ID, Double ROUND_NUMBER, Double ROUND_PICK, Double OVERALL_PICK, Double TEAM_ID, String PLAYER_NAME, String SEASON, String DRAFT_TYPE, String TEAM_CITY, String TEAM_NAME, String TEAM_ABBREVIATION, String ORGANIZATION, String ORGANIZATION_TYPE) {
+    public DraftPick(int PERSON_ID, int ROUND_NUMBER, int ROUND_PICK, int OVERALL_PICK, int TEAM_ID, String PLAYER_NAME, String SEASON, String DRAFT_TYPE, String TEAM_CITY, String TEAM_NAME, String TEAM_ABBREVIATION, String ORGANIZATION, String ORGANIZATION_TYPE) {
         this.PERSON_ID = PERSON_ID;
         this.ROUND_NUMBER = ROUND_NUMBER;
         this.ROUND_PICK = ROUND_PICK;
@@ -27,43 +27,43 @@ public class DraftPick implements Serializable {
     public DraftPick() {
     }
 
-    public Double getPERSON_ID() {
+    public int getPERSON_ID() {
         return PERSON_ID;
     }
 
-    public void setPERSON_ID(Double PERSON_ID) {
+    public void setPERSON_ID(int PERSON_ID) {
         this.PERSON_ID = PERSON_ID;
     }
 
-    public Double getROUND_NUMBER() {
+    public int getROUND_NUMBER() {
         return ROUND_NUMBER;
     }
 
-    public void setROUND_NUMBER(Double ROUND_NUMBER) {
+    public void setROUND_NUMBER(int ROUND_NUMBER) {
         this.ROUND_NUMBER = ROUND_NUMBER;
     }
 
-    public Double getROUND_PICK() {
+    public int getROUND_PICK() {
         return ROUND_PICK;
     }
 
-    public void setROUND_PICK(Double ROUND_PICK) {
+    public void setROUND_PICK(int ROUND_PICK) {
         this.ROUND_PICK = ROUND_PICK;
     }
 
-    public Double getOVERALL_PICK() {
+    public int getOVERALL_PICK() {
         return OVERALL_PICK;
     }
 
-    public void setOVERALL_PICK(Double OVERALL_PICK) {
+    public void setOVERALL_PICK(int OVERALL_PICK) {
         this.OVERALL_PICK = OVERALL_PICK;
     }
 
-    public Double getTEAM_ID() {
+    public int getTEAM_ID() {
         return TEAM_ID;
     }
 
-    public void setTEAM_ID(Double TEAM_ID) {
+    public void setTEAM_ID(int TEAM_ID) {
         this.TEAM_ID = TEAM_ID;
     }
